Move `get_email_summary()` to `RepositoryUtils.php`
